+++ onlyoffice_odoo/tests/__init__.py
@@ -6,5 +6,6 @@ from . import test_controllers
 from . import test_file_utils
 from . import test_jwt_utils
 from . import test_models
+from . import test_network_utils
 from . import test_url_utils
 from . import test_validation_utils